Slowly getting rid of the syntax guide
The concrete syntax should really be exclusively in the popovers
I think.  In this way, I think I can probably keep the reference
very compact and also well organized.  I might even be able to
get rid of the tabs at the top (in which case everything will
be visible with no more than one click.