Comment on issue #4
Although I haven't added coverage of the synchronous features in
Chapter 16 of 3.3, I added a comment where I say this covers
Modelica 3.3 indicating that the synchronous features are not
yet covered.

I'll leave #4 open (since I'd like to add synchronous stuff at
some point), but hopefully this at least makes it clear
the synchronous stuff is understood to be missing.