I added some caveats about screen size
This is mostly in reference to #5.  At some point, I might experiment
around with the responsive features in Bootstrap.  But, frankly, it was so
difficult to organize everything with a fixed layout that I have a hard
time seeing how the responsive features could do anything but make a
mess of the content.  I'd like to try, but this is definitely a
'version 2.0' issue.  For now, people will just have to leave with
the (not unreasonable) screen size limitations.